// packages/engine/src/templates/shared/DefaultInput.d.ts
export declare enum DefaultInput {
    PRIMARY = 0,
    SECONDARY = 1,
    FORWARD = 2,
    BACKWARD = 3,
    UP = 4,
    DOWN = 5,
    LEFT = 6,
    RIGHT = 7,
    INTERACT = 8,
    CROUCH = 9,
    JUMP = 10,
    WALK = 11,
    RUN = 12,
    SPRINT = 13,
    SNEAK = 14,
    SCREENXY = 15,
    SCREENXY_START = 16,
    ROTATION_START = 17,
    MOVEMENT_PLAYERONE = 18,
    LOOKTURN_PLAYERONE = 19,
    MOVEMENT_PLAYERTWO = 20,
    LOOKTURN_PLAYERTWO = 21,
    ALTERNATE = 22,
    MOUSE_MOVEMENT = 23,
    SWITCH_CAR = 24,
    POINTER_LOCK = 25,
    CAMERA_SCROLL = 26,
    SWITCH_CAMERA = 27,
    SWITCH_SHOULDER_SIDE = 28,
    LOCKING_CAMERA = 29,
    POINTER1_POSITION = 30,
    POINTER2_POSITION = 31,
    GAMEPAD_STICK_LEFT = 32,
    GAMEPAD_STICK_RIGHT = 33,
    FACE_EXPRESSION = 34,
    FACE_EXPRESSION_HAPPY = 35,
    FACE_EXPRESSION_SAD = 36
}